refactor abstract operator into discreteoperator
