+++ netket/operator/__init__.py
@@ -14,6 +14,7 @@
 
 from ._abstract_operator import AbstractOperator
 
+from ._discrete_operator import DiscreteOperator
 from ._local_operator import LocalOperator
 from ._graph_operator import GraphOperator
 from ._pauli_strings import PauliStrings
